Only Design check only checks m_top.
The previous version of Design check would go through all models, starting with
the models of standard cells and checking m_top last.

The Model check first does a bidir check which is irrelevant for all models, and
then it starts verifying the connectivity of m_nets, which is from the model, against
b_nets, which is from m_top. That makes no sense.

It's clear the Model check method is written from the point of view of checking a
user provided design, not the standard cell models.

Since the blif parser ensures that there is only one m_top, Design check now only
checks m_top and that is that.